Reuse existing serialization code and move helpers into Node
